requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Admin
gunicorn
psycopg2-binary
requests
python-dotenv
cloudinary
Pillow


